ZeroShot + get_framework import.
